refactor skills route detail pane
